Fix many issues parsing and writing TFS monster abilities (#18)
These changes fix issues #6, #8, and #14 (both the melee issue and combat damage type issue)

Other improvements beyond the reported issues

    TFS XML: Fixed parsing of condition causing spells, melee attack name, duration, animations, effects, and area information for all spells
    TFS RevScriptSys: Writing condition causing spells, duration, area information, outfit changing spells, animations, and effects for all, and combat type
    TFSXML->TFS RevScriptSys: Spells from files now in some cases but needs more testing and checking overall
