+++ app/MonsterConverterInterface/Enums/Animation.cs
@@ -1,0 +1,61 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MonsterConverterInterface.MonsterTypes
+{
+    public enum Animation
+    {
+        None = 0,
+        Spear,
+        Bolt,
+        Arrow,
+        Fire,
+        Energy,
+        PoisonArrow,
+        BurstArrow,
+        ThrowingStar,
+        ThrowingKnife,
+        SmallStone,
+        Death,
+        LargeRock,
+        Snowball,
+        PowerBolt,
+        Poison,
+        InfernalBolt,
+        HuntingSpear,
+        EnchantedSpear,
+        RedStar,
+        GreenStar,
+        RoyalSpear,
+        SniperArrow,
+        OnyxArrow,
+        PiercingBolt,
+        WhirlwindSword,
+        WhirlwindAxe,
+        WhirlwindClub,
+        EtherealSpear,
+        Ice,
+        Earth,
+        Holy,
+        SuddenDeath,
+        FlashArrow,
+        FlammingArrow,
+        ShiverArrow,
+        EnergyBall,
+        SmallIce,
+        SmallHoly,
+        SmallEarth,
+        EarthArrow,
+        Explosion,
+        Cake,
+        TarsalArrow = 44,
+        VortexBolt = 45,
+        PrismaticBolt,
+        CrystallineArrow,
+        DrillBolt,
+        EnvenomedArrow,
+        GloothSpear,
+        SimpleArrow
+    }
+}
