+++ app/MonsterConverterInterface/Enums/Blood.cs
@@ -1,0 +1,15 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MonsterConverterInterface.MonsterTypes
+{
+    public enum Blood
+    {
+        blood,
+        venom,
+        undead,
+        fire,
+        energy
+    }
+}
