+++ app/MonsterConverterInterface/Enums/SoundLevel.cs
@@ -1,0 +1,13 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Text;
+
+namespace MonsterConverterInterface.MonsterTypes
+{
+    public enum SoundLevel
+    {
+        Whisper,
+        Say,
+        Yell
+    }
+}
