+++ app/MonsterConverterInterface/Enums/SpellCategory.cs
